+++ go/integTest/shared_commands_test.go
@@ -6590,7 +6590,7 @@ func (suite *GlideTestSuite) TestDumpRestore() {
 		assert.Equal(t, int64(1), deletedCount)
 		result_test1, err := client.Restore(key, int64(0), resultDump.Value())
 		assert.Nil(suite.T(), err)
-		assert.Equal(suite.T(), "OK", result_test1.Value())
+		assert.Equal(suite.T(), "OK", result_test1)
 		resultGetRestoreKey, err := client.Get(key)
 		assert.Nil(t, err)
 		assert.Equal(t, value, resultGetRestoreKey.Value())
@@ -6621,7 +6621,7 @@ func (suite *GlideTestSuite) TestRestoreWithOptions() {
 		optsReplace := options.NewRestoreOptions().SetReplace()
 		result_test1, err := client.RestoreWithOptions(key, int64(0), resultDump.Value(), *optsReplace)
 		assert.Nil(suite.T(), err)
-		assert.Equal(suite.T(), "OK", result_test1.Value())
+		assert.Equal(suite.T(), "OK", result_test1)
 		resultGetRestoreKey, err := client.Get(key)
 		assert.Nil(t, err)
 		assert.Equal(t, value, resultGetRestoreKey.Value())
@@ -6633,7 +6633,7 @@ func (suite *GlideTestSuite) TestRestoreWithOptions() {
 		opts_test2 := options.NewRestoreOptions().SetABSTTL()
 		result_test2, err := client.RestoreWithOptions(key, int64(0), resultDump.Value(), *opts_test2)
 		assert.Nil(suite.T(), err)
-		assert.Equal(suite.T(), "OK", result_test2.Value())
+		assert.Equal(suite.T(), "OK", result_test2)
 		resultGet_test2, err := client.Get(key)
 		assert.Nil(t, err)
 		assert.Equal(t, value, resultGet_test2.Value())
@@ -6645,7 +6645,7 @@ func (suite *GlideTestSuite) TestRestoreWithOptions() {
 		opts_test3 := options.NewRestoreOptions().SetEviction(options.FREQ, 10)
 		result_test3, err := client.RestoreWithOptions(key, int64(0), resultDump.Value(), *opts_test3)
 		assert.Nil(suite.T(), err)
-		assert.Equal(suite.T(), "OK", result_test3.Value())
+		assert.Equal(suite.T(), "OK", result_test3)
 		resultGet_test3, err := client.Get(key)
 		assert.Nil(t, err)
 		assert.Equal(t, value, resultGet_test3.Value())
@@ -6657,7 +6657,7 @@ func (suite *GlideTestSuite) TestRestoreWithOptions() {
 		opts_test4 := options.NewRestoreOptions().SetEviction(options.IDLETIME, 10)
 		result_test4, err := client.RestoreWithOptions(key, int64(0), resultDump.Value(), *opts_test4)
 		assert.Nil(suite.T(), err)
-		assert.Equal(suite.T(), "OK", result_test4.Value())
+		assert.Equal(suite.T(), "OK", result_test4)
 		resultGet_test4, err := client.Get(key)
 		assert.Nil(t, err)
 		assert.Equal(t, value, resultGet_test4.Value())
@@ -10129,5 +10129,213 @@ func (suite *GlideTestSuite) TestGeoSearchStore() {
 		_, err = client.GeoSearchStore(destinationKey, key3, searchOrigin, *boxShape)
 		assert.Error(suite.T(), err)
 		assert.IsType(suite.T(), &errors.RequestError{}, err)
+	})
+}
+
+func (suite *GlideTestSuite) TestBZPopMax() {
+	suite.SkipIfServerVersionLowerThanBy("7.0.0")
+
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key1 := "{key}-1" + uuid.NewString()
+
+		res1, err := client.BZPopMax([]string{key1}, float64(0.1))
+		assert.Nil(suite.T(), err)
+		assert.True(suite.T(), res1.IsNil())
+
+		membersScoreMap := map[string]float64{
+			"one":   1.0,
+			"two":   2.0,
+			"three": 3.0,
+		}
+
+		res2, err := client.ZAdd(key1, membersScoreMap)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), int64(3), res2)
+
+		res3, err := client.BZPopMax([]string{key1}, float64(0.1))
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), api.KeyWithMemberAndScore{Key: key1, Member: "three", Score: 3.0}, res3.Value())
+	})
+}
+
+func (suite *GlideTestSuite) TestZMPop() {
+	suite.SkipIfServerVersionLowerThanBy("7.0.0")
+
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key1 := "{key}-1" + uuid.NewString()
+		key2 := "{key}-2" + uuid.NewString()
+		key3 := "{key}-3" + uuid.NewString()
+
+		res1, err := client.ZMPop([]string{key1}, options.MIN)
+		assert.Nil(suite.T(), err)
+		assert.True(suite.T(), res1.IsNil())
+
+		membersScoreMap := map[string]float64{
+			"one":   1.0,
+			"two":   2.0,
+			"three": 3.0,
+		}
+		res2, err := client.ZAdd(key1, membersScoreMap)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), int64(3), res2)
+
+		res3, err := client.ZAdd(key2, map[string]float64{
+			"four": 4.0,
+			"five": 5.0,
+		})
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), int64(2), res3)
+
+		// Pop minimum value from key1
+		res4, err := client.ZMPop([]string{key1}, options.MIN)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), key1, res4.Value().Key)
+		assert.ElementsMatch(
+			suite.T(),
+			[]api.MemberAndScore{
+				{Member: "one", Score: 1.0},
+			},
+			res4.Value().MembersAndScores,
+		)
+
+		// Pop maximum value from key2
+		res5, err := client.ZMPop([]string{key2}, options.MAX)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), key2, res5.Value().Key)
+		assert.ElementsMatch(
+			suite.T(),
+			[]api.MemberAndScore{
+				{Member: "five", Score: 5.0},
+			},
+			res5.Value().MembersAndScores,
+		)
+
+		// pop from an empty key3
+		res6, err := client.ZMPop([]string{key3}, options.MIN)
+		assert.Nil(suite.T(), err)
+		assert.True(suite.T(), res6.IsNil())
+	})
+}
+
+func (suite *GlideTestSuite) TestZMPopWithOptions() {
+	suite.SkipIfServerVersionLowerThanBy("7.0.0")
+
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key1 := "{key}-1" + uuid.NewString()
+		key2 := "{key}-2" + uuid.NewString()
+		key3 := "{key}-3" + uuid.NewString()
+
+		opts := *options.NewZPopOptions().SetCount(2)
+
+		res1, err := client.ZMPopWithOptions([]string{key1}, options.MIN, opts)
+		assert.Nil(suite.T(), err)
+		assert.True(suite.T(), res1.IsNil())
+
+		membersScoreMap := map[string]float64{
+			"one":   1.0,
+			"two":   2.0,
+			"three": 3.0,
+			"four":  4.0,
+		}
+		res2, err := client.ZAdd(key1, membersScoreMap)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), int64(4), res2)
+
+		res3, err := client.ZAdd(key2, map[string]float64{
+			"a": 10.0,
+			"b": 20.0,
+		})
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), int64(2), res3)
+
+		res4, err := client.ZMPopWithOptions([]string{key1}, options.MIN, opts)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), key1, res4.Value().Key)
+		assert.ElementsMatch(
+			suite.T(),
+			[]api.MemberAndScore{
+				{Member: "one", Score: 1.0},
+				{Member: "two", Score: 2.0},
+			},
+			res4.Value().MembersAndScores,
+		)
+
+		opts10 := *options.NewZPopOptions().SetCount(10)
+		res5, err := client.ZMPopWithOptions([]string{key1}, options.MIN, opts10)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), key1, res5.Value().Key)
+		assert.ElementsMatch(
+			suite.T(),
+			[]api.MemberAndScore{
+				{Member: "three", Score: 3.0},
+				{Member: "four", Score: 4.0},
+			},
+			res5.Value().MembersAndScores,
+		)
+
+		opts1 := *options.NewZPopOptions().SetCount(1)
+		res6, err := client.ZMPopWithOptions([]string{key2}, options.MAX, opts1)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), key2, res6.Value().Key)
+		assert.ElementsMatch(
+			suite.T(),
+			[]api.MemberAndScore{
+				{Member: "b", Score: 20.0},
+			},
+			res6.Value().MembersAndScores,
+		)
+
+		res7, err := client.ZMPopWithOptions([]string{key3}, options.MIN, opts1)
+		assert.Nil(suite.T(), err)
+		assert.True(suite.T(), res7.IsNil())
+	})
+}
+
+func (suite *GlideTestSuite) TestInvokeScriptWithoutRoute() {
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key1 := uuid.NewString()
+		key2 := uuid.NewString()
+
+		// Test a script that returns a string without keys and args.
+		script1 := options.NewScript("return 'Hello'")
+		response1, err := client.InvokeScript(*script1)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "Hello", response1)
+
+		// Test script that sets a key with value.
+		script2 := options.NewScript("return redis.call('SET', KEYS[1], ARGV[1])")
+
+		// Create Script options for setting key1
+		scriptOptions := options.NewScriptOptions()
+		scriptOptions.WithKeys([]string{key1}).WithArgs([]string{"value1"})
+		setResponse, err := client.InvokeScriptWithOptions(*script2, *scriptOptions)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "OK", setResponse)
+
+		// Set another key, key2 with the same script
+		scriptOptions2 := options.NewScriptOptions()
+		scriptOptions2.WithKeys([]string{key2}).WithArgs([]string{"value2"})
+		setResponse2, err := client.InvokeScriptWithOptions(*script2, *scriptOptions2)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "OK", setResponse2)
+		script2.Close()
+
+		// Test script that gets a key's value
+		script3 := options.NewScript("return redis.call('GET', KEYS[1])")
+
+		// Create ClusterScriptOptions for getting key1
+		scriptOptions3 := options.NewScriptOptions()
+		scriptOptions3.WithKeys([]string{key1})
+		getResponse1, err := client.InvokeScriptWithOptions(*script3, *scriptOptions3)
+		assert.Nil(suite.T(), err)
+		assert.Equal(suite.T(), "value1", getResponse1)
+
+		// Get another key's value
+		scriptOptions4 := options.NewScriptOptions()
+		scriptOptions4.WithKeys([]string{key2})
+		getResponse2, err := client.InvokeScriptWithOptions(*script3, *scriptOptions4)
+		assert.Equal(suite.T(), "value2", getResponse2)
+		assert.Nil(suite.T(), err)
+		script3.Close()
 	})
 }